feat(shop-the-look): componentize selection and config UI
This commit marks the completion of Phase 2 of the "Shop the Look" refactoring plan.

This phase focuses on decomposing the monolithic UI into smaller, more manageable components for the initial user interaction steps.

The primary changes include:
- Creating a new `components/shop_the_look/` directory.
- Creating a `config_panel.py` component to house the UI and logic from the "Config" tab.
- Creating a `model_selection.py` component for the "Choose a Model" screen.
- Creating a `look_selection.py` component for the "Choose a Look" screen.
- Updating `pages/shop_the_look.py` to import and use these new, focused components, removing the old code.
