+++ experiments/veo-app/models/shop_the_look_workflow.py
@@ -13,8 +13,10 @@
 # limitations under the License.
 
 """Workflow for the Shop the Look page."""
+
 import csv
 import datetime
+
 import mesop as me
 from google.cloud import firestore
 
@@ -28,11 +30,189 @@ from models.shop_the_look_models import (
     CatalogRecord,
     ModelRecord,
 )
-from state.state import AppState
 from state.shop_the_look_state import PageState
+from state.state import AppState
 
 config = Default()
 db = FirebaseClient(database_id=config.GENMEDIA_FIREBASE_DB).get_client()
+
+def model_on_delete(e: me.ClickEvent):
+     state = me.state(PageState)
+     file_to_delete = e.key.split("/")[-1]
+     print(f"deleting {file_to_delete}")
+     state.current_status = f"Deleting model {file_to_delete}"
+     try:
+         doc_ref = db.collection(config.GENMEDIA_VTO_MODEL_COLLECTION_NAME).document(
+             file_to_delete
+         )
+
+         doc_ref.delete()
+         state.models = load_model_data()
+         state.current_status = ""
+         yield
+     except:
+         print(f"Model data  delete failure: {file_to_delete} cannot be stored")
+
+def article_on_delete(e: me.ClickEvent):
+    state = me.state(PageState)
+    file_to_delete = e.key.split("/")[-1]
+    print(f"deleting {file_to_delete}")
+    state.current_status = f"Deleting article {file_to_delete}"
+    try:
+        doc_ref = db.collection(config.GENMEDIA_VTO_CATALOG_COLLECTION_NAME).document(
+            file_to_delete
+        )
+        doc_ref.delete()
+        load_article_data()
+        state.current_status = ""
+        yield
+    except:
+        print(f"Model data  delete failure: {file_to_delete} cannot be stored")
+
+
+def get_csv_headers(csv_reader):
+    """
+    Retrieves a list of header names from a CSV file.
+
+    Args:
+        filepath (str): The path to the CSV file.
+
+    Returns:
+        list: A list containing the header names, or an empty list if the file is empty or an error occurs.
+    """
+    try:
+        header = next(csv_reader)
+        return header
+    except Exception as e:
+        print(f"An error occurred: {e}")
+        return []
+
+
+def on_click_upload_models(e: me.UploadEvent):
+    """Upload image to GCS"""
+    state = me.state(PageState)
+    state.reference_model_file = e.file
+    contents = e.file.getvalue()
+    destination_blob_name = store_to_gcs(
+        "uploads", e.file.name, e.file.mime_type, contents
+    )
+
+    state.reference_model_file_gs_uri = f"gs://{destination_blob_name}"
+
+    print(
+        f"{destination_blob_name} with contents len {len(contents)} of type {e.file.mime_type} uploaded to {config.GENMEDIA_BUCKET}."
+    )
+
+    csv_file = download_from_gcs_as_string(
+        f"gs://{config.GENMEDIA_BUCKET}/uploads/{e.file.name}"
+    )
+
+    cf = [row.decode("utf-8") for row in csv_file.split(b"\n") if row]
+    cf = csv.reader(cf, delimiter=",")
+
+    required_fields = [
+        "model_group",
+        "model_id",
+        "model_name",
+        "model_description",
+        "model_view",
+        "primary_view",
+        "model_image",
+    ]
+
+    headers = get_csv_headers(cf)
+
+    for c in required_fields:  # ie. ["batch", "department"]
+        if c not in headers:
+            print(f"Missing CSV header for {c}")
+            return
+
+    current_datetime = datetime.datetime.now()
+
+    for row in cf:
+        try:
+            # TODO mapping object instead of row[]
+            doc_ref = db.collection(config.GENMEDIA_VTO_MODEL_COLLECTION_NAME).document(
+                f"{row[1]}_{row[4]}"
+            )
+            doc_ref.set(
+                {
+                    "model_group": row[0],
+                    "model_id": row[1],
+                    "model_name": row[2],
+                    "model_description": row[3],
+                    "model_view": row[4],
+                    "primary_view": row[5],
+                    "model_image": row[6],
+                    "timestamp": current_datetime,  # alt: firestore.SERVER_TIMESTAMP
+                }
+            )
+        except:
+            print(f"{row[2]} cannot be converted")
+
+
+def on_click_upload_catalog(e: me.UploadEvent):
+    """Upload image to GCS"""
+    state = me.state(PageState)
+    state.reference_catalog_file = e.file
+    contents = e.file.getvalue()
+    destination_blob_name = store_to_gcs(
+        "uploads", e.file.name, e.file.mime_type, contents
+    )
+
+    state.reference_catalog_file_gs_uri = f"gs://{destination_blob_name}"
+
+    print(
+        f"{destination_blob_name} with contents len {len(contents)} of type {e.file.mime_type} uploaded to {config.GENMEDIA_BUCKET}."
+    )
+
+    csv_file = download_from_gcs_as_string(
+        f"gs://{config.GENMEDIA_BUCKET}/uploads/{e.file.name}"
+    )
+
+    cf = [row.decode("utf-8") for row in csv_file.split(b"\n") if row]
+    cf = csv.reader(cf, delimiter=",")
+
+    required_fields = [
+        "item_id",
+        "look_id",
+        "article_type",
+        "article_color",
+        "model_group",
+        "description",
+        "image_view",
+        "try_on_order",
+    ]
+
+    headers = get_csv_headers(cf)
+
+    for c in required_fields:  # ie. ["batch", "department"]
+        if c not in headers:
+            print(f"Missing CSV header for {c}")
+            return
+
+    current_datetime = datetime.datetime.now()
+
+    for row in cf:
+        try:
+            doc_ref = db.collection(
+                config.GENMEDIA_VTO_CATALOG_COLLECTION_NAME
+            ).document(f"{row[1]}_{row[2]}")
+            doc_ref.set(
+                {
+                    "item_id": row[0],
+                    "look_id": int(row[1]),
+                    "article_type": row[2],
+                    "article_color": row[3],
+                    "model_group": row[4],
+                    "description": row[5],
+                    "image_view": row[6],
+                    "try_on_order": row[7],
+                    "timestamp": current_datetime,  # alt: firestore.SERVER_TIMESTAMP
+                }
+            )
+        except:
+            print(f"{row[2]} cannot be converted")
 
 
 def load_model_data(limit: int = 50):
@@ -107,6 +287,7 @@ def get_model_records(model_id):
             model_records.append(m)
     return model_records
 
+
 def store_model_data(file_path):
     state = me.state(PageState)
     app_state = me.state(AppState)
@@ -149,38 +330,3 @@ def store_article_data(file_path, article_category):
     }
 
     update_time, doc_ref = doc_ref.add(new_doc_data)
-
-def article_on_delete(e: me.ClickEvent):
-    state = me.state(PageState)
-    file_to_delete = e.key.split("/")[-1]
-    print(f"deleting {file_to_delete}")
-    state.current_status = f"Deleting article {file_to_delete}"
-    try:
-        doc_ref = db.collection(config.GENMEDIA_VTO_CATALOG_COLLECTION_NAME).document(
-            file_to_delete
-        )
-
-        doc_ref.delete()
-        load_article_data()
-        state.current_status = ""
-        yield
-    except:
-        print(f"Model data  delete failure: {file_to_delete} cannot be stored")
-
-
-def model_on_delete(e: me.ClickEvent):
-    state = me.state(PageState)
-    file_to_delete = e.key.split("/")[-1]
-    print(f"deleting {file_to_delete}")
-    state.current_status = f"Deleting model {file_to_delete}"
-    try:
-        doc_ref = db.collection(config.GENMEDIA_VTO_MODEL_COLLECTION_NAME).document(
-            file_to_delete
-        )
-
-        doc_ref.delete()
-        state.models = load_model_data()
-        state.current_status = ""
-        yield
-    except:
-        print(f"Model data  delete failure: {file_to_delete} cannot be stored")
