+++ experiments/veo-app/pages/shop_the_look.py
@@ -40,6 +40,9 @@ from components.page_scaffold import (
     page_frame,
     page_scaffold,
 )
+from components.shop_the_look.config_panel import config_panel
+from components.shop_the_look.look_selection import look_selection
+from components.shop_the_look.model_selection import model_selection
 from components.tab_nav import Tab, tab_group
 from config.default import Default
 from models.gemini import (
@@ -68,7 +71,7 @@ config = Default()
 def tab_config():
     state = me.state(PageState)
     state.models = shop_the_look_workflow.load_model_data()
-    vto_enterprise_config()
+    config_panel()
 
 
 @me.component
@@ -80,9 +83,9 @@ def tab_stl():
         state.models = shop_the_look_workflow.load_model_data()
         shop_the_look_workflow.load_article_data()
     if state.reference_image_gcs_model is None or not state.reference_image_gcs_model:
-        stl_model_select()
+        model_selection()
     elif not state.look or state.look == 0:
-        stl_look_select()
+        look_selection()
     else:
         stl_result()
 
@@ -243,175 +246,6 @@ def on_click_veo(e: me.ClickEvent):  # pylint: disable=unused-argument
     state.current_status = ""
     yield
     print("Cut! That's a wrap!")
-
-
-def stl_model_select():
-    state = me.state(PageState)
-
-    with me.box(
-        style=me.Style(
-            display="flex",
-            flex_direction="column",
-            gap=10,
-            align_items="center",
-            width="100%",
-        )
-    ):
-        me.text(
-            text="Choose a Model",
-            type="headline-2",
-            # style=me.Style(width="100%", text_align="center"),
-        )
-    with me.box():
-        with me.box(
-            style=me.Style(
-                height="100%",
-                width="100%",
-                display="flex",
-                flex_direction="row",
-                flex_wrap="wrap",
-            )
-        ):
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="left",
-                    # margin=me.Margin(left=10),
-                )
-            ):
-                with me.box(
-                    style=me.Style(
-                        position="relative",
-                        height="100%",
-                        margin=me.Margin(left=7, top=7),
-                        cursor="pointer",
-                    ),
-                    key="model",
-                    # on_click=on_click_upload_image,
-                ):
-                    me.uploader(
-                        label="",
-                        accepted_file_types=["image/jpeg", "image/png"],
-                        on_upload=on_click_upload_image,
-                        type="flat",
-                        color="primary",
-                        style=me.Style(
-                            position="relative",
-                            # margin=me.Margin(left=7, top=7),
-                            cursor="pointer",
-                            width="200px",  # Consider using max_width or responsive units
-                            height="200px",
-                            object_fit="cover",
-                            border_radius="5px",
-                            box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                            background="#FFFFFF",
-                        ),
-                        key="model",
-                        multiple=True,
-                    )
-                    # me.box(
-                    #     style=me.Style(
-                    #         width="200px",  # Consider using max_width or responsive units
-                    #         height="200px",
-                    #         object_fit="cover",
-                    #         border_radius="5px",
-                    #         box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                    #         background="#FFFFFF",
-                    #     ),
-                    # )
-                    (
-                        me.icon(
-                            "add",
-                            style=me.Style(
-                                color="black",
-                                position="absolute",
-                                top="35px",
-                                left="35px",
-                                width="35px",
-                                height="35px",
-                                font_size="35px",
-                            ),  # Consider using max_width or responsive units
-                        ),
-                    )
-                    (
-                        me.icon(
-                            "person_outline",
-                            style=me.Style(
-                                color="black",
-                                position="absolute",
-                                top="50px",
-                                left="55px",
-                                width="100px",
-                                height="100px",
-                                font_size="100px",
-                            ),  # Consider using max_width or responsive units
-                        ),
-                    )
-                    (
-                        me.text(
-                            text="Add Model",
-                            style=me.Style(
-                                position="absolute",
-                                # width="100%",
-                                text_align="center",
-                                bottom="45px",
-                                left="55px",
-                                font_size="20px",
-                            ),
-                        ),
-                    )
-
-            for model in state.models:
-                with me.box(
-                    style=me.Style(
-                        display="flex",
-                        flex_direction="row",
-                        gap=5,
-                        align_items="left",
-                        # margin=me.Margin(left=10),
-                    )
-                ):
-                    with me.box(
-                        key=f"{model.model_image}",
-                        style=me.Style(
-                            position="relative",
-                            height="100%",
-                            margin=me.Margin(left=7, top=7),
-                            cursor="pointer",
-                        ),
-                        on_click=on_model_click,
-                    ):
-                        # me.icon(
-                        #     "delete_forever",
-                        #     style=me.Style(
-                        #         color="black",
-                        #         position="absolute",
-                        #         top="5px",
-                        #         right="5px",
-                        #         width="20px",
-                        #         height="20px",
-                        #         font_size=("20px"),
-                        #     ),
-                        # )
-                        me.image(
-                            src=model.model_image.replace(
-                                "gs://",
-                                "https://storage.mtls.cloud.google.com/",
-                            ),
-                            style=me.Style(
-                                # width="200px",  # Consider using max_width or responsive units
-                                # height="200px",
-                                object_fit="cover",
-                                border_radius="5px",
-                                box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                                max_height="200px",
-                                height="auto",
-                            ),
-                        )
-
-    return me
 
 
 def stl_result():
@@ -1000,449 +834,6 @@ def on_click_manual_retry(e: me.ClickEvent):
     yield from on_click_vto_look(new_event)
 
 
-def article_on_click(e: me.ClickEvent):
-    state = me.state(PageState)
-    selected_type = e.key.split("_")[-1]
-    selected_id = e.key.split("_")[-2]
-
-    selected = False
-    for item in state.articles:
-        print(f"comparing {item.item_id} to {selected_id}")
-        if item.item_id == selected_id:
-            item.selected = not item.selected
-            selected = item.selected
-
-    for item in state.articles:
-        if item.item_id != selected_id:
-            if selected_type == "shoe" and item.article_type == "shoe":
-                item.available_to_select = not selected
-                item.selected = False
-            elif selected_type == "dress" and item.article_type in [
-                "dress",
-                "top",
-                "bottom",
-            ]:
-                item.available_to_select = not selected
-                item.selected = False
-            elif selected_type == "top" and item.article_type in ["dress", "top"]:
-                item.available_to_select = not selected
-                item.selected = False
-            elif selected_type == "bottom" and item.article_type in ["dress", "bottom"]:
-                item.available_to_select = not selected
-                item.selected = False
-
-    yield
-
-
-def stl_look_select():
-    state = me.state(PageState)
-
-    with me.box(
-        style=me.Style(
-            display="flex",
-            flex_direction="column",
-            gap=10,
-            align_items="center",
-            width="100%",
-        )
-    ):
-        me.text(
-            text="Choose a Look",
-            type="headline-2",
-            # style=me.Style(width="100%", text_align="center"),
-        )
-    with me.box():
-        with me.box(
-            style=me.Style(
-                height="100%",
-                width="100%",
-                display="flex",
-                flex_direction="row",
-                flex_wrap="wrap",
-            )
-        ):
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="left",
-                    # margin=me.Margin(left=10),
-                )
-            ):
-                with me.box(
-                    style=me.Style(
-                        position="relative",
-                        height="100%",
-                        cursor="pointer",
-                    ),
-                    key="apparel",
-                    # on_click=on_click_upload_image,
-                ):
-                    me.uploader(
-                        label="",
-                        accepted_file_types=["image/jpeg", "image/png"],
-                        on_upload=on_click_upload_image,
-                        type="flat",
-                        color="primary",
-                        style=me.Style(
-                            position="relative",
-                            cursor="pointer",
-                            width="150px",  # Consider using max_width or responsive units
-                            height="150px",
-                            object_fit="cover",
-                            border_radius="5px",
-                            box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                            background="#FFFFFF",
-                            margin=me.Margin(left=10, top=10),
-                        ),
-                        key="top",
-                        multiple=True,
-                    )
-                    (
-                        me.icon(
-                            "add",
-                            style=me.Style(
-                                color="black",
-                                position="absolute",
-                                top="20px",
-                                left="30px",
-                                width="30px",
-                                height="30px",
-                                font_size="30px",
-                            ),  # Consider using max_width or responsive units
-                        ),
-                    )
-                    (
-                        me.icon(
-                            "apparel",
-                            style=me.Style(
-                                color="black",
-                                position="absolute",
-                                top="45px",
-                                left="55px",
-                                width="50px",
-                                height="50px",
-                                font_size="50px",
-                            ),  # Consider using max_width or responsive units
-                        ),
-                    )
-                    (
-                        me.text(
-                            text="Add Top",
-                            style=me.Style(
-                                position="absolute",
-                                # width="100%",
-                                text_align="center",
-                                bottom="35px",
-                                left="45px",
-                                font_size="20px",
-                            ),
-                        ),
-                    )
-            with me.box(
-                style=me.Style(
-                    position="relative",
-                    height="100%",
-                    cursor="pointer",
-                ),
-                key="apparel",
-                # on_click=on_click_upload_image,
-            ):
-                me.uploader(
-                    label="",
-                    accepted_file_types=["image/jpeg", "image/png"],
-                    on_upload=on_click_upload_image,
-                    type="flat",
-                    color="primary",
-                    style=me.Style(
-                        position="relative",
-                        cursor="pointer",
-                        width="150px",  # Consider using max_width or responsive units
-                        height="150px",
-                        object_fit="cover",
-                        border_radius="5px",
-                        box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                        background="#FFFFFF",
-                        margin=me.Margin(left=10, top=10),
-                    ),
-                    key="bottom",
-                    multiple=True,
-                )
-                (
-                    me.icon(
-                        "add",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="20px",
-                            left="30px",
-                            width="30px",
-                            height="30px",
-                            font_size="30px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                )
-                (
-                    me.icon(
-                        "styler",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="45px",
-                            left="55px",
-                            width="50px",
-                            height="50px",
-                            font_size="50px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                )
-                (
-                    me.text(
-                        text="Add Bottom",
-                        style=me.Style(
-                            position="absolute",
-                            # width="100%",
-                            text_align="center",
-                            bottom="35px",
-                            left="30px",
-                            font_size="20px",
-                        ),
-                    ),
-                )
-
-            with me.box(
-                style=me.Style(
-                    position="relative",
-                    height="100%",
-                    cursor="pointer",
-                ),
-                key="apparel",
-                # on_click=on_click_upload_image,
-            ):
-                me.uploader(
-                    label="",
-                    accepted_file_types=["image/jpeg", "image/png"],
-                    on_upload=on_click_upload_image,
-                    type="flat",
-                    color="primary",
-                    style=me.Style(
-                        position="relative",
-                        cursor="pointer",
-                        width="150px",  # Consider using max_width or responsive units
-                        height="150px",
-                        object_fit="cover",
-                        border_radius="5px",
-                        box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                        background="#FFFFFF",
-                        margin=me.Margin(left=10, top=10),
-                    ),
-                    key="dress",
-                    multiple=True,
-                )
-                (
-                    me.icon(
-                        "add",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="20px",
-                            left="30px",
-                            width="30px",
-                            height="30px",
-                            font_size="30px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                )
-                (
-                    me.icon(
-                        "girl",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="35px",
-                            left="45px",
-                            width="70px",
-                            height="70px",
-                            font_size="70px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                )
-                (
-                    me.text(
-                        text="Add Dress",
-                        style=me.Style(
-                            position="absolute",
-                            # width="100%",
-                            text_align="center",
-                            bottom="35px",
-                            left="35px",
-                            font_size="20px",
-                        ),
-                    ),
-                )
-
-            with me.box(
-                style=me.Style(
-                    position="relative",
-                    height="100%",
-                    cursor="pointer",
-                ),
-                key="apparel",
-                # on_click=on_click_upload_image,
-            ):
-                me.uploader(
-                    label="",
-                    accepted_file_types=["image/jpeg", "image/png"],
-                    on_upload=on_click_upload_image,
-                    type="flat",
-                    color="primary",
-                    style=me.Style(
-                        position="relative",
-                        cursor="pointer",
-                        width="150px",  # Consider using max_width or responsive units
-                        height="150px",
-                        object_fit="cover",
-                        border_radius="5px",
-                        box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                        background="#FFFFFF",
-                        margin=me.Margin(left=10, top=10),
-                    ),
-                    key="shoe",
-                    multiple=True,
-                )
-                (
-                    me.icon(
-                        "add",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="20px",
-                            left="30px",
-                            width="30px",
-                            height="30px",
-                            font_size="30px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                )
-                (
-                    me.icon(
-                        "steps",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="45px",
-                            left="55px",
-                            width="50px",
-                            height="50px",
-                            font_size="50px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                )
-                (
-                    me.text(
-                        text="Add Shoe",
-                        style=me.Style(
-                            position="absolute",
-                            # width="100%",
-                            text_align="center",
-                            bottom="35px",
-                            left="35px",
-                            font_size="20px",
-                        ),
-                    ),
-                )
-
-            for item in state.articles:
-                with me.box(
-                    style=me.Style(
-                        display="flex",
-                        flex_direction="row",
-                        gap=5,
-                        align_items="left",
-                        # margin=me.Margin(left=10),
-                    )
-                ):
-                    img = item.clothing_image.replace(
-                        "gs://",
-                        "https://storage.mtls.cloud.google.com/",
-                    )
-                    print(f"item.clothing_image {item.clothing_image}")
-                    with me.box(
-                        key=f"{item.item_id}_{item.article_type}",
-                        style=me.Style(
-                            position="relative",
-                            height="100%",
-                            # display=("block" if item.available_to_select else "none"),
-                            margin=me.Margin(left=10, top=10),
-                            cursor="pointer",
-                        ),
-                        on_click=article_on_click,
-                    ):
-                        me.icon(
-                            "check_circle",
-                            style=me.Style(
-                                color=("green" if item.selected else "#666"),
-                                position="absolute",
-                                top="1px",
-                                left="5px",
-                                opacity=("1" if item.selected else ".1"),
-                            ),
-                        )
-
-                        me.image(
-                            src=img,
-                            style=me.Style(
-                                width="150px",  # Consider using max_width or responsive units
-                                height="150px",
-                                object_fit="cover",
-                                border_radius="5px",
-                                box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                                opacity=(
-                                    "1"
-                                    if (item.available_to_select or item.selected)
-                                    else ".3"
-                                ),
-                            ),
-                        )
-    with me.box(
-        style=me.Style(
-            display="flex",
-            flex_direction="row",
-            gap=5,
-            align_items="right",
-        ),
-    ):
-        me.button(
-            "Continue",
-            type="flat",
-            style=me.Style(
-                width=75,
-                height=40,
-                border_radius=5,
-                cursor="pointer",
-                margin=me.Margin(left=5, top=10),
-            ),
-            on_click=on_continue_click,
-        )
-
-
-def on_model_click(e: me.ClickEvent):
-    state = me.state(PageState)
-    state.reference_image_gcs_model = e.key
-    state.before_image_uri = e.key
-    print(f"state.before_image_uri {state.before_image_uri}")
-    for m in state.models:
-        if m.model_image in e.key:
-            state.selected_model = m
-
-
-def on_continue_click(e: me.ClickEvent):
-    state = me.state(PageState)
-    print(e)
-    state.look = 2
-
-
 def on_look_button_click(e: me.ClickEvent):
     state = me.state(PageState)
     print(e)
@@ -1456,366 +847,7 @@ def on_config_generate_alt_views(event: me.CheckboxChangeEvent):
     state.generate_alternate_views = event.checked
 
 
-def on_config_generate_videos(event: me.CheckboxChangeEvent):
-    state = me.state(PageState)
-    state.generate_video = event.checked
-
-
-def on_config_upload_everyone(event: me.CheckboxChangeEvent):
-    state = me.state(PageState)
-    state.upload_everyone = event.checked
-
-
-def on_veo_version_change(e: me.SelectSelectionChangeEvent):
-    s = me.state(PageState)
-    s.veo_model = e.value
-
-
-def on_sample_count_change(e: me.SelectSelectionChangeEvent):
-    s = me.state(PageState)
-    s.vto_sample_count = e.value
-
-
-def max_retry_change(e: me.SelectSelectionChangeEvent):
-    s = me.state(PageState)
-    s.max_retry = e.value
-
-
-def on_blur_veo_prompt(e: me.InputBlurEvent):
-    """Veo prompt blur event"""
-    me.state(PageState).veo_prompt_input = e.value
-
-
 # End Config Events
-
-
-def vto_enterprise_config():
-    state = me.state(PageState)
-    app_state = me.state(AppState)
-
-    with me.box():
-        with me.box(
-            style=me.Style(
-                height="100%",
-                width="100%",
-                display="flex",
-                flex_direction="row",
-                flex_wrap="wrap",
-            )
-        ):
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                )
-            ):
-                me.text(
-                    text="VTO",
-                    type="headline-6",
-                )
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                )
-            ):
-                me.checkbox(
-                    "Upload model/clothing for all access?",
-                    checked=state.upload_everyone,
-                    on_change=on_config_upload_everyone,
-                )
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                )
-            ):
-                me.select(
-                    label="VTO Sample Count",
-                    options=[
-                        me.SelectOption(label="1 image", value="1"),
-                        me.SelectOption(label="2 images", value="2"),
-                        me.SelectOption(label="3 images", value="3"),
-                        me.SelectOption(label="4 images", value="4"),
-                    ],
-                    on_selection_change=on_sample_count_change,
-                    style=me.Style(width=180),
-                    multiple=False,
-                    appearance="outline",
-                    value=state.vto_sample_count,
-                )
-                me.select(
-                    label="Max Auto Retry",
-                    options=[
-                        me.SelectOption(label="1 times", value="1"),
-                        me.SelectOption(label="2 times", value="2"),
-                        me.SelectOption(label="3 times", value="3"),
-                        me.SelectOption(label="4 times", value="4"),
-                    ],
-                    on_selection_change=max_retry_change,
-                    style=me.Style(width=180),
-                    multiple=False,
-                    appearance="outline",
-                    value=state.max_retry,
-                )
-
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                )
-            ):
-                me.text(
-                    text="VEO",
-                    type="headline-6",
-                )
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                )
-            ):
-                me.select(
-                    label="Version",
-                    options=[
-                        me.SelectOption(label="Veo 3", value="3.0"),
-                        me.SelectOption(label="Veo 2", value="2.0"),
-                    ],
-                    on_selection_change=on_veo_version_change,
-                    style=me.Style(width=140),
-                    multiple=False,
-                    appearance="outline",
-                    value=state.veo_model,
-                )
-                me.checkbox(
-                    "Generate videos (in addition)",
-                    checked=state.generate_video,
-                    on_change=on_config_generate_videos,
-                )
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                )
-            ):
-                me.text(
-                    text="Base Prompt",
-                    style=me.Style(
-                        font_size="12px",
-                    ),
-                )
-            me.native_textarea(
-                autosize=True,
-                min_rows=10,
-                max_rows=13,
-                placeholder="Base prompt",
-                style=me.Style(
-                    padding=me.Padding(top=16, left=16),
-                    background=me.theme_var("secondary-container"),
-                    outline="none",
-                    width="1200px",
-                    overflow_y="auto",
-                    border=me.Border.all(
-                        me.BorderSide(style="none"),
-                    ),
-                    color=me.theme_var("foreground"),
-                    flex_grow=1,
-                ),
-                on_blur=on_blur_veo_prompt,
-                # key=str(state.veo_prompt_textarea_key),
-                value=state.veo_prompt_input,
-            )
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                    margin=me.Margin(top=10),
-                )
-            ):
-                me.text(
-                    text="Apparel",
-                    type="headline-6",
-                )
-            if state.articles:
-                for item in state.articles:
-                    with me.box(
-                        style=me.Style(
-                            display="flex",
-                            flex_direction="row",
-                            gap=5,
-                            align_items="left",
-                        )
-                    ):
-                        img = item.clothing_image.replace(
-                            "gs://",
-                            "https://storage.mtls.cloud.google.com/",
-                        )
-                        print(f"item.clothing_image {item.clothing_image}")
-                        with me.box(
-                            key=f"{item.item_id}-{item.article_type}",
-                            style=me.Style(
-                                position="relative",
-                                height="100%",
-                                margin=me.Margin(left=10, top=10),
-                            ),
-                        ):
-                            with me.box(
-                                key=f"{img}",
-                                on_click=shop_the_look_workflow.article_on_delete,
-                                style=me.Style(
-                                    cursor="pointer",
-                                ),
-                            ):
-                                me.icon(
-                                    "delete",
-                                    style=me.Style(
-                                        color="red",
-                                        position="absolute",
-                                        top="5px",
-                                        left="5px",
-                                        height="30px",
-                                        width="30px",
-                                        font_size="30px",
-                                    ),
-                                )
-
-                            me.image(
-                                src=img,
-                                style=me.Style(
-                                    width="150px",  # Consider using max_width or responsive units
-                                    height="150px",
-                                    object_fit="cover",
-                                    border_radius="5px",
-                                    box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                                ),
-                            )
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                )
-            ):
-                for item in state.reference_image_uri_clothing:
-                    with me.box(
-                        style=me.Style(
-                            display="flex",
-                            flex_direction="row",
-                            gap=5,
-                            align_items="left",
-                        )
-                    ):
-                        with me.box(
-                            key=f"{item}",
-                            style=me.Style(
-                                position="relative",
-                                height="100%",
-                                margin=me.Margin(left=7, top=7),
-                                cursor="pointer",
-                            ),
-                        ):
-                            me.image(
-                                src=f"{item}",
-                                style=me.Style(
-                                    width="200px",  # Consider using max_width or responsive units
-                                    height="200px",
-                                    object_fit="cover",
-                                    border_radius="5px",
-                                    box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                                ),
-                            )
-            with me.box(
-                style=me.Style(
-                    display="flex",
-                    flex_direction="row",
-                    gap=5,
-                    align_items="top",
-                    width="100%",
-                    margin=me.Margin(top=10),
-                )
-            ):
-                me.text(
-                    text="Models",
-                    type="headline-6",
-                )
-            print(state.models)
-            if state.models:
-                for model in state.models:
-                    with me.box(
-                        style=me.Style(
-                            display="flex",
-                            flex_direction="row",
-                            gap=5,
-                            align_items="left",
-                        )
-                    ):
-                        img = model.model_image.replace(
-                            "gs://",
-                            "https://storage.mtls.cloud.google.com/",
-                        )
-                        print(f"model.model_image {model.model_image}")
-                        with me.box(
-                            style=me.Style(
-                                position="relative",
-                                height="100%",
-                                margin=me.Margin(left=10, top=10),
-                            ),
-                        ):
-                            with me.box(
-                                key=f"{img}",
-                                on_click=shop_the_look_workflow.model_on_delete,
-                                style=me.Style(
-                                    cursor="pointer",
-                                ),
-                            ):
-                                me.icon(
-                                    "delete",
-                                    style=me.Style(
-                                        color="red",
-                                        position="absolute",
-                                        top="5px",
-                                        left="5px",
-                                        height="30px",
-                                        width="30px",
-                                        font_size="30px",
-                                    ),
-                                )
-
-                            me.image(
-                                src=img,
-                                style=me.Style(
-                                    width="150px",  # Consider using max_width or responsive units
-                                    height="150px",
-                                    object_fit="cover",
-                                    border_radius="5px",
-                                    box_shadow="0 2px 4px rgba(0,0,0,0.1)",  # Use theme shadow if available
-                                ),
-                            )
-
-    return me
 
 
 def build_tab_nav():
@@ -1857,62 +889,6 @@ def on_tab_click(e: me.ClickEvent):
 
     next(on_click_clear_reference_image())
     state.selected_tab_index = int(tab_index)
-
-
-def on_click_upload_image(e: me.UploadEvent):
-    """Upload image to GCS"""
-    state = me.state(PageState)
-
-    for i, file in enumerate(e.files):
-        state.current_status = f"Uploading {i + 1} of {len(e.files)}"
-        state.reference_image_file_clothing = e.file
-        yield
-        filename_uuid = str(uuid.uuid4())
-        file_ext = file.name.split(".")[-1]
-        filename = f"{filename_uuid}.{file_ext}"
-
-        if e.key == "model":
-            file_path = f"gs://{config.GENMEDIA_BUCKET}/uploads/models/{filename}"
-            gcs_url = store_to_gcs(
-                "uploads/models",
-                filename.lower(),
-                file.mime_type,
-                file.getvalue(),
-            )
-            shop_the_look_workflow.store_model_data(file_path)
-        else:
-            file_path = (
-                f"gs://{config.GENMEDIA_BUCKET}/uploads/apparel/{e.key}/{filename}"
-            )
-            gcs_url = store_to_gcs(
-                f"uploads/apparel/{e.key}",
-                filename.lower(),
-                file.mime_type,
-                file.getvalue(),
-            )
-            state.reference_image_gcs_clothing.append(f"{gcs_url}")
-            state.reference_image_uri_clothing.append(
-                gcs_url.replace(
-                    "gs://",
-                    "https://storage.mtls.cloud.google.com/",
-                )
-            )
-            article_type = gcs_url.split("/")[-2]
-            shop_the_look_workflow.store_article_data(file_path, article_type)
-
-        print(
-            f"{gcs_url} of type {e.file.mime_type} uploaded to {config.GENMEDIA_BUCKET}."
-        )
-
-        state.current_status = ""
-        yield
-
-    if e.key == "model":
-        state.models = shop_the_look_workflow.load_model_data()
-        yield
-    else:
-        shop_the_look_workflow.load_article_data()
-        yield
 
 
 def on_click_vto_look(e: me.ClickEvent):  # pylint: disable=unused-argument
